create and update: common component ResetToMyLocation
