+++ src/components/maptalks/MaptalksCheckpoint.jsx
@@ -7,6 +7,7 @@ import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { blue, yellow } from "@mui/material/colors";
 
 import { calculateAreaOfRectangle } from "./calculateAreaOfRectangle";
+import ResetToMyLocation from "./ResetToMyLocation";
 
 const MaptalksCheckpoint = () => {
   const mapRef = useRef(null); // Ref to store the map DOM element
@@ -166,7 +167,7 @@ const MaptalksCheckpoint = () => {
         baseLayer: new maptalks.TileLayer("base", {
           urlTemplate: "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
           subdomains: ["a", "b", "c"],
-          attribution: "© OpenStreetMap contributors",
+          attribution: "Vehicle Tracking",
         }),
       });
 
@@ -228,17 +229,6 @@ const MaptalksCheckpoint = () => {
       secondary: yellow,
     },
   });
-
-  const resetToMyLocation = () => {
-    const savedCoords = sessionStorage.getItem("userCoordinates");
-    if (savedCoords) {
-      const userCenter = JSON.parse(savedCoords);
-      mapInstance.current.setCenter(userCenter);
-      mapInstance.current.setZoom(16); // Reset zoom to a closer view
-    } else {
-      alert("User location not available.");
-    }
-  };
 
   // Handle name change
   const handleNameChange = (index, newName) => {
@@ -302,29 +292,7 @@ const MaptalksCheckpoint = () => {
                 <span className="sentencebutton">Reset Shape</span>
               </Button>
 
-              <Button
-                variant="outlined"
-                onClick={resetToMyLocation}
-                color="primary"
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="28"
-                  height="28"
-                  viewBox="0 0 24 24"
-                  fill="none"
-                  stroke="red"
-                  strokeWidth="0.75"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  className="lucide lucide-map-pin-check-inside"
-                >
-                  <path d="M20 10c0 4.993-5.539 10.193-7.399 11.799a1 1 0 0 1-1.202 0C9.539 20.193 4 14.993 4 10a8 8 0 0 1 16 0" />
-                  <path d="m9 10 2 2 4-4" />
-                </svg>
-                &nbsp; &nbsp;{" "}
-                <span className="sentencebutton">Reset To My Location</span>
-              </Button>
+              <ResetToMyLocation mapInstance={mapInstance} />
             </Stack>
             <div className="card bottomradius bigcard everythingCenter">
               <div className="card2 dynamicheight bottomradius bigcard routeCords">
